split the tests -into two files
