+++ tests/test_quotes.py
@@ -1,7 +1,7 @@
 from datetime import date
 
 from app.database.models import Quote
-from app.internal.quotes import load_quotes, daily_quotes
+from app.internal.quotes import daily_quotes
 
 DATE = date(2021, 1, 1)
 DATE2 = date(2021, 1, 2)
@@ -14,26 +14,6 @@ def insert_quotes(session):
     session.add(quote)
     session.add(quote2)
     session.commit()
-
-
-# Tests for loading the quotes in the db:
-def test_load_daily_quotes(session):
-    load_quotes.load_daily_quotes(session)
-    assert session.query(Quote).count() > 0
-
-
-def test_load_daily_quotes_with_json_valueerror(mocker, session):
-    mock_json_load = mocker.patch('json.load')
-    mock_json_load.side_effect = ValueError
-    load_quotes.load_daily_quotes(session)
-    assert session.query(Quote).count() == 0
-
-
-def test_quotes_not_load_twice_to_db(session):
-    load_quotes.load_daily_quotes(session)
-    first_quotes_amount = session.query(Quote).count()
-    load_quotes.load_daily_quotes(session)
-    assert first_quotes_amount == session.query(Quote).count()
 
 
 # Tests for providing a daily-quote from the db:
